move MonteCarloBarostat from utils.py to equilibrium.py
